Feature: ABC for DataCatalog (#674)
* Created ABC for DataCatalog

* Adjusted import paths for DataCatalog ABC

* Created ABC for DataCatalog

* Created ABC for DataCatalog

* Updated test for changes to DataCatalog

* Added imports to reduce changes impact

* :sparkles: added an ABC for DataCatalog

* Updated tests new DataCatalog definitions
