+++ nautilus_trader/config/common.py
@@ -27,7 +27,7 @@ from pydantic import validator
 
 from nautilus_trader.common import Environment
 from nautilus_trader.core.correctness import PyCondition
-from nautilus_trader.persistence.catalog import DataCatalog
+from nautilus_trader.persistence.catalog.parquet import ParquetDataCatalog
 
 
 def resolve_path(path: str):
@@ -220,11 +220,11 @@ class StreamingConfig(NautilusConfig):
         return fsspec.filesystem(protocol=self.fs_protocol, **(self.fs_storage_options or {}))
 
     @classmethod
-    def from_catalog(cls, catalog: DataCatalog, **kwargs):
+    def from_catalog(cls, catalog: ParquetDataCatalog, **kwargs):
         return cls(catalog_path=str(catalog.path), fs_protocol=catalog.fs.protocol, **kwargs)
 
-    def as_catalog(self) -> DataCatalog:
-        return DataCatalog(
+    def as_catalog(self) -> ParquetDataCatalog:
+        return ParquetDataCatalog(
             path=self.catalog_path,
             fs_protocol=self.fs_protocol,
             fs_storage_options=self.fs_storage_options,
