+++ nautilus_trader/persistence/catalog/__init__.py
@@ -12,30 +12,13 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
-
-from nautilus_trader.persistence.catalog.base import BaseDataCatalog
-from nautilus_trader.persistence.external.core import write_objects
-
-
-# TODO (bm)
+from .base import BaseDataCatalog
+from .parquet import ParquetDataCatalog
+from .parquet import resolve_path
 
 
-def create_temp_table(func):
-    """Make a temporary copy of any parquet dataset class called by `write_tables`"""
-
-    def inner(*args, **kwargs):
-        try:
-            return func(*args, **kwargs)
-        except Exception:
-            # Restore old table
-            print()
-
-    return inner
-
-
-write_objects = create_temp_table(write_objects)
-
-
-def migrate(catalog: BaseDataCatalog, version_from: str, version_to: str):
-    """Migrate the `catalog` between versions `version_from` and `version_to`"""
-    pass
+__all__ = (
+    "BaseDataCatalog",
+    "ParquetDataCatalog",
+    "resolve_path",
+)
