+++ nautilus_trader/persistence/catalog/base.py
@@ -13,18 +13,14 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
-import os
-import pathlib
-import platform
+from abc import ABC
+from abc import ABCMeta
+from abc import abstractclassmethod
+from abc import abstractmethod
 from typing import Callable, Dict, List, Optional, Union
 
-import fsspec
 import pandas as pd
 import pyarrow as pa
-import pyarrow.dataset as ds
-import pyarrow.parquet as pq
-from fsspec.utils import infer_storage_options
-from pyarrow import ArrowInvalid
 
 from nautilus_trader.core.inspect import is_nautilus_class
 from nautilus_trader.model.data.bar import Bar
@@ -39,58 +35,30 @@ from nautilus_trader.model.orderbook.data import OrderBookData
 from nautilus_trader.persistence.base import Singleton
 from nautilus_trader.persistence.external.metadata import load_mappings
 from nautilus_trader.serialization.arrow.serializer import ParquetSerializer
-from nautilus_trader.serialization.arrow.serializer import list_schemas
 from nautilus_trader.serialization.arrow.util import GENERIC_DATA_PREFIX
-from nautilus_trader.serialization.arrow.util import camel_to_snake_case
-from nautilus_trader.serialization.arrow.util import class_to_filename
-from nautilus_trader.serialization.arrow.util import clean_key
 from nautilus_trader.serialization.arrow.util import dict_of_lists_to_list_of_dicts
 
 
-class DataCatalog(metaclass=Singleton):
-    """
-    Provides a queryable data catalog.
+class _CombinedMeta(Singleton, ABCMeta):
+    pass
 
-    Parameters
-    ----------
-    path : str
-        The root path for this data catalog. Must exist and must be an absolute path.
-    fs_protocol : str, default 'file'
-        The fsspec filesystem protocol to use.
-    fs_storage_options : Dict, optional
-        The fs storage options.
+
+class BaseDataCatalog(ABC, metaclass=_CombinedMeta):
+    """
+    Provides a abstract base class for a queryable data catalog.
     """
 
-    def __init__(
-        self,
-        path: str,
-        fs_protocol: str = "file",
-        fs_storage_options: Optional[Dict] = None,
-    ):
-        self.fs_protocol = fs_protocol
-        self.fs_storage_options = fs_storage_options or {}
-        self.fs: fsspec.AbstractFileSystem = fsspec.filesystem(
-            self.fs_protocol, **self.fs_storage_options
-        )
-        self.path: pathlib.Path = pathlib.Path(path)
-
-    @classmethod
+    @abstractclassmethod
     def from_env(cls):
-        return cls.from_uri(uri=os.path.join(os.environ["NAUTILUS_PATH"], "catalog"))
+        raise NotImplementedError
 
-    @classmethod
+    @abstractclassmethod
     def from_uri(cls, uri):
-        if "://" not in uri:
-            # Assume a local path
-            uri = "file://" + uri
-        parsed = infer_storage_options(uri)
-        path = parsed.pop("path")
-        protocol = parsed.pop("protocol")
-        storage_options = parsed.copy()
-        return cls(path=path, fs_protocol=protocol, fs_storage_options=storage_options)
+        raise NotImplementedError
 
     # -- QUERIES -----------------------------------------------------------------------------------
 
+    @abstractmethod
     def _query(
         self,
         cls: type,
@@ -107,38 +75,7 @@ class DataCatalog(metaclass=Singleton):
         projections: Optional[Dict] = None,
         **kwargs,
     ):
-        filters = [filter_expr] if filter_expr is not None else []
-        if instrument_ids is not None:
-            if not isinstance(instrument_ids, list):
-                instrument_ids = [instrument_ids]
-            if clean_instrument_keys:
-                instrument_ids = list(set(map(clean_key, instrument_ids)))
-            filters.append(ds.field(instrument_id_column).cast("string").isin(instrument_ids))
-        if start is not None:
-            filters.append(ds.field(ts_column) >= int(pd.Timestamp(start).to_datetime64()))
-        if end is not None:
-            filters.append(ds.field(ts_column) <= int(pd.Timestamp(end).to_datetime64()))
-
-        full_path = str(self._make_path(cls=cls))
-        if not (self.fs.exists(full_path) or self.fs.isdir(full_path)):
-            if raise_on_empty:
-                raise FileNotFoundError(f"protocol={self.fs.protocol}, path={full_path}")
-            else:
-                return pd.DataFrame() if as_dataframe else None
-
-        dataset = ds.dataset(full_path, partitioning="hive", filesystem=self.fs)
-        table_kwargs = table_kwargs or {}
-        if projections:
-            projected = {**{c: ds.field(c) for c in dataset.schema.names}, **projections}
-            table_kwargs.update(columns=projected)
-        table = dataset.to_table(filter=combine_filters(*filters), **(table_kwargs or {}))
-        mappings = self.load_inverse_mappings(path=full_path)
-        if as_dataframe:
-            return self._handle_table_dataframe(
-                table=table, mappings=mappings, raise_on_empty=raise_on_empty, **kwargs
-            )
-        else:
-            return self._handle_table_nautilus(table=table, cls=cls, mappings=mappings)
+        raise NotImplementedError
 
     def load_inverse_mappings(self, path):
         mappings = load_mappings(fs=self.fs, path=path)
@@ -189,10 +126,6 @@ class DataCatalog(metaclass=Singleton):
         data = ParquetSerializer.deserialize(cls=cls, chunk=dicts)
         return data
 
-    def _make_path(self, cls: type) -> str:
-        path: pathlib.Path = self.path / "data" / f"{class_to_filename(cls=cls)}.parquet"
-        return str(resolve_path(path=path, fs=self.fs))
-
     def query(
         self,
         cls: type,
@@ -222,6 +155,7 @@ class DataCatalog(metaclass=Singleton):
             **kwargs,
         )
 
+    @abstractmethod
     def _query_subclasses(
         self,
         base_cls: type,
@@ -230,34 +164,7 @@ class DataCatalog(metaclass=Singleton):
         as_nautilus: bool = False,
         **kwargs,
     ):
-        subclasses = [base_cls] + base_cls.__subclasses__()
-
-        dfs = []
-        for cls in subclasses:
-            try:
-                df = self._query(
-                    cls=cls,
-                    filter_expr=filter_expr,
-                    instrument_ids=instrument_ids,
-                    raise_on_empty=False,
-                    as_dataframe=not as_nautilus,
-                    **kwargs,
-                )
-                dfs.append(df)
-            except ArrowInvalid as ex:
-                # If we're using a `filter_expr` here, there's a good chance
-                # this error is using a filter that is specific to one set of
-                # instruments and not to others, so we ignore it (if not; raise).
-                if filter_expr is not None:
-                    continue
-                else:
-                    raise ex
-
-        if not as_nautilus:
-            return pd.concat([df for df in dfs if df is not None])
-        else:
-            objects = [o for objs in filter(None, dfs) for o in objs]
-            return objects
+        raise NotImplementedError
 
     def instruments(
         self,
@@ -394,9 +301,9 @@ class DataCatalog(metaclass=Singleton):
             return [GenericData(data_type=DataType(cls), data=d) for d in data]
         return data
 
+    @abstractmethod
     def list_data_types(self):
-        glob_path = resolve_path(self.path / "data" / "*.parquet", fs=self.fs)
-        return [pathlib.Path(p).stem for p in self.fs.glob(glob_path)]
+        raise NotImplementedError
 
     def list_generic_data_types(self):
         data_types = self.list_data_types()
@@ -406,101 +313,18 @@ class DataCatalog(metaclass=Singleton):
             if n.startswith(GENERIC_DATA_PREFIX)
         ]
 
-    def list_partitions(self, cls_type: type):
-        assert isinstance(cls_type, type), "`cls_type` should be type, i.e. TradeTick"
-        name = class_to_filename(cls_type)
-        dataset = pq.ParquetDataset(
-            resolve_path(self.path / f"{name}.parquet", fs=self.fs), filesystem=self.fs
-        )
-        partitions = {}
-        for level in dataset.partitions.levels:
-            partitions[level.name] = level.keys
-        return partitions
-
+    @abstractmethod
     def list_backtests(self) -> List[str]:
-        glob = resolve_path(self.path / "backtest" / "*.feather", fs=self.fs)
-        return [p.stem for p in map(pathlib.Path, self.fs.glob(glob))]
+        raise NotImplementedError
 
+    @abstractmethod
     def list_live_runs(self) -> List[str]:
-        glob = resolve_path(self.path / "live" / "*.feather", fs=self.fs)
-        return [p.stem for p in map(pathlib.Path, self.fs.glob(glob))]
+        raise NotImplementedError
 
+    @abstractmethod
     def read_live_run(self, live_run_id: str, **kwargs):
-        return self._read_feather(kind="live", run_id=live_run_id, **kwargs)
+        raise NotImplementedError
 
+    @abstractmethod
     def read_backtest(self, backtest_run_id: str, **kwargs):
-        return self._read_feather(kind="backtest", run_id=backtest_run_id, **kwargs)
-
-    def _read_feather(self, kind: str, run_id: str, raise_on_failed_deserialize: bool = False):
-        class_mapping: Dict[str, type] = {class_to_filename(cls): cls for cls in list_schemas()}
-        data = {}
-        glob_path = resolve_path(self.path / kind / f"{run_id}.feather" / "*.feather", fs=self.fs)
-        for path in [p for p in self.fs.glob(glob_path)]:
-            cls_name = camel_to_snake_case(pathlib.Path(path).stem).replace("__", "_")
-            df = read_feather_file(path=path, fs=self.fs)
-            if df is None:
-                print(f"No data for {cls_name}")
-                continue
-            # Apply post read fixes
-            try:
-                objs = self._handle_table_nautilus(
-                    table=df, cls=class_mapping[cls_name], mappings={}
-                )
-                data[cls_name] = objs
-            except Exception as ex:
-                if raise_on_failed_deserialize:
-                    raise
-                print(f"Failed to deserialize {cls_name}: {ex}")
-        return sorted(sum(data.values(), list()), key=lambda x: x.ts_init)
-
-
-def read_feather_file(path: str, fs: fsspec.AbstractFileSystem = None):
-    fs = fs or fsspec.filesystem("file")
-    if not fs.exists(path):
-        return
-    try:
-        with fs.open(path) as f:
-            reader = pa.ipc.open_stream(f)
-            return reader.read_pandas()
-    except (pa.ArrowInvalid, FileNotFoundError):
-        return
-
-
-def combine_filters(*filters):
-    filters = tuple(x for x in filters if x is not None)
-    if len(filters) == 0:
-        return
-    elif len(filters) == 1:
-        return filters[0]
-    else:
-        expr = filters[0]
-        for f in filters[1:]:
-            expr = expr & f
-        return expr
-
-
-def _should_use_windows_paths(fs: fsspec.filesystem) -> bool:
-    """
-    Pathlib will try and use windows style paths even when a fsspec.filesystem does not (memory, s3, etc).
-
-    We need to determine the case when we should use windows paths, which is when we are on windows and using a
-    fsspec.filesystem that is local.
-
-    """
-    from fsspec.implementations.local import LocalFileSystem
-
-    try:
-        from fsspec.implementations.smb import SMBFileSystem
-    except ImportError:
-        SMBFileSystem = LocalFileSystem
-
-    is_windows = platform.system() == "Windows"
-    is_windows_local_fs = isinstance(fs, (LocalFileSystem, SMBFileSystem))
-    return is_windows and is_windows_local_fs
-
-
-def resolve_path(path: pathlib.Path, fs: fsspec.filesystem) -> str:
-    if _should_use_windows_paths(fs=fs):
-        return str(path)
-    else:
-        return path.as_posix()
+        raise NotImplementedError
