+++ tests/integration_tests/adapters/betfair/test_betfair_persistence.py
@@ -17,7 +17,7 @@ import fsspec
 import pytest
 
 from nautilus_trader.adapters.betfair.data_types import BSPOrderBookDelta
-from nautilus_trader.persistence.catalog import DataCatalog
+from nautilus_trader.persistence.catalog.parquet import ParquetDataCatalog
 from nautilus_trader.persistence.external.core import RawFile
 from nautilus_trader.persistence.external.core import process_raw_file
 from tests.integration_tests.adapters.betfair.test_kit import BetfairTestStubs
@@ -28,7 +28,7 @@ from tests.test_kit.mocks.data import data_catalog_setup
 class TestBetfairPersistence:
     def setup(self):
         data_catalog_setup()
-        self.catalog = DataCatalog.from_env()
+        self.catalog = ParquetDataCatalog.from_env()
         self.fs = self.catalog.fs
         self.reader = BetfairTestStubs.betfair_reader()
