+++ tests/unit_tests/backtest/test_backtest_engine.py
@@ -52,7 +52,7 @@ from nautilus_trader.model.orderbook.data import Order
 from nautilus_trader.model.orderbook.data import OrderBookDelta
 from nautilus_trader.model.orderbook.data import OrderBookDeltas
 from nautilus_trader.model.orderbook.data import OrderBookSnapshot
-from nautilus_trader.persistence.catalog import DataCatalog
+from nautilus_trader.persistence.catalog.parquet import ParquetDataCatalog
 from nautilus_trader.trading.strategy import Strategy
 from tests.test_kit.stubs import MyData
 from tests.test_kit.stubs.component import TestComponentStubs
@@ -159,7 +159,7 @@ class TestBacktestEngine:
     def test_persistence_files_cleaned_up(self):
         # Arrange
         temp_dir = tempfile.mkdtemp()
-        catalog = DataCatalog(
+        catalog = ParquetDataCatalog(
             path=str(temp_dir),
             fs_protocol="file",
         )
